+++ packages/nuget/pack-nuget.sh
@@ -8,7 +8,12 @@ fi
 
 mkdir -p $nugetfeedpath
 
+echo "Fetching Nuget executable"
+wget https://dist.nuget.org/win-x86-commandline/latest/nuget.exe
+echo "Packing Package"
+mono nuget.exe pack .
+
 echo "Deleting existing package"
-mono nuget.exe delete ${NUGET_PACKAGE_NAME} ${CMAKE_PROJECT_VERSION} -Source $nugetfeedpath -NonInteractive
+mono nuget.exe delete ${NUGET_PACKAGE_NAME}.${NUGET_ARCH} ${CMAKE_PROJECT_VERSION} -Source $nugetfeedpath -NonInteractive
 echo "Installing new package"
 mono nuget.exe init . $nugetfeedpath