+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.4] - 2023-02-13
+### Fixed
+- Added LICENSE and README to NuGet Package
+
+
+## [1.0.3] - 2023-02-13
+### Fixed
+- Fixed NuGet Package Generation
+
+## [1.0.2] - 2023-02-12
+### Changed
+- Generating separate NuGet Packages per architecture
+
 ## [1.0.0] - 2023-02-06
 ### Added
 - Coordinate Utilities
@@ -135,6 +148,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Conversion Functions
   - Legendre Spherical Harmonics Functions (For Magnetic Calculations)
 
+[1.0.4]: https://github.com/997R8V10/aviation-calc-util/compare/v1.0.3...v1.0.4
+[1.0.3]: https://github.com/997R8V10/aviation-calc-util/compare/v1.0.2...v1.0.3
+[1.0.2]: https://github.com/997R8V10/aviation-calc-util/compare/v1.0.0...v1.0.2
 [1.0.0]: https://github.com/997R8V10/aviation-calc-util/compare/v0.3.6...v1.0.0
 [0.3.6]: https://github.com/997R8V10/aviation-calc-util/compare/v0.3.5...v0.3.6
 [0.3.5]: https://github.com/997R8V10/aviation-calc-util/compare/v0.3.4...v0.3.5